pyproject.toml
[project]
name = "Gdylagents"
version = "0.1.0"
description = "我的智能Agents集合"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=1.0.2",
    "langchain-anthropic>=1.0.3",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "deepagents>=0.2.6",
    "python-dotenv>=1.0.0",
    "langgraph-cli[inmem]>=0.1.55",
    "duckduckgo-search>=8.1.1",
    "lxml>=6.0.2",
]

[tool.uv]
override-dependencies = [
    "nbconvert>=7.17.0",
    "protobuf>=6.33.5",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["research_agent","tester_agent"]

[tool.setuptools.package-data]
"*" = ["py.typed"]
